Voice/requirements_vosk.txt
vosk
sounddevice
numpy
websocket-server
python-dotenv
pyaudio